+++ src/pages/Lending/Pool/components/Repay/RepayInput.tsx
@@ -39,7 +39,7 @@ import { bn, bnOrZero } from 'lib/bignumber/bignumber'
 import { getMaybeCompositeAssetSymbol } from 'lib/mixpanel/helpers'
 import { getMixPanel } from 'lib/mixpanel/mixPanelSingleton'
 import { MixPanelEvent } from 'lib/mixpanel/types'
-import { useThorAllowance } from 'lib/utils/thorchain/hooks/useIsThorApprovalNeeded'
+import { useThorAllowance } from 'lib/utils/thorchain/hooks/useThorAllowance'
 import type { LendingQuoteClose } from 'lib/utils/thorchain/lending/types'
 import { useLendingQuoteCloseQuery } from 'pages/Lending/hooks/useLendingCloseQuery'
 import { useLendingPositionData } from 'pages/Lending/hooks/useLendingPositionData'
